requirements.txt
opencv-contrib-python
opencv-python==4.6.0.66
albumentations==1.3.0
tqdm==4.64.0
tornado==6.2
bbox-visualizer==0.1.0
pybboxes==0.1.5
scikit-image==0.19.3
Pillow==9.1.1
matplotlib
rich==12.6.0